feat!: WDM representation and WDM Whittle noise model
Introduce the WDM time-frequency representation using pywavelet as
a backend (new dependency).

Also introduce the WDMData container and the WDMWhittleNoise noise model.
It is a Gaussian noise model whose covariance matrix is diagonal across
time-frequency pixels but nondiagonal across TDI channels.

Much of the functionality in WDMWhittleNoise are convenience methods to
treat the nonzero elements of the covariance matrix, called an SDM
(spectral density matrix) in the code.

BREAKING CHANGE: this is more advanced than the old WDMWhittleNoise in tlt-wdm
and the interface is incompatible with that one.